V8

+++ src/CohortData.cs
@@ -297,7 +297,7 @@ namespace Landis.Library.PnETCohorts
         /// <summary>
         /// The cohort's ANPP
         /// </summary>
-        public int ANPP;
+        public double ANPP;
 
         /// <summary>
         /// Initializes a new instance.
@@ -359,7 +359,7 @@ namespace Landis.Library.PnETCohorts
         /// The age of the cohort.
         /// <param name="totalBioamss">
         /// The biomamss of the cohort
-        public CohortData(ushort age, float totalBiomass, int totalANPP, ISpecies species, bool cohortStacking)
+        public CohortData(ushort age, float totalBiomass, double totalANPP, ISpecies species, bool cohortStacking)
         {            
             this.AdjFolN = new float[Globals.IMAX];
             this.adjFolN = 0; ;
